% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/super_func.R
\name{super_func}
\alias{super_func}
\title{sum between two vectors}
\usage{
super_func(x, y)
}
\arguments{
\item{x}{numeric vector}

\item{y}{numeric vector}
}
\value{
numeric vector
}
\description{
sum between two vectors
}
\examples{
super_func(2:5, 6:9)
}
